refactor: define `ahc` module
